services/ingestor_sentinel/requirements.txt
httpx==0.27.*
asyncpg==0.29.*
redis==5.0.*
pyyaml==6.0.*
numpy==1.26.*
